Streamline CI jobs
Add net8.0 as target framework.

Remove the `doctests` workflow, since all tests inside `tests/Doc` are
executed anyway during the `integration` workflow.

Separate the concurrency groups between workflows, to avoid them
cancelling each other during one CI run. This seems to be a known issue:
https://github.com/orgs/community/discussions/41518

Update the release workflow to use dotnet 8, given that we have it as a
target now.

Print the dotnet version and available SDKs and runtimes in some
workflows, for clarity.

Update the `reusable` workflow to build for a specific target framework,
and to use a specific version of the dotnet SDK when building and
running tests. This is done via a generated `global.json` file. The goal
is to ensure that the source code can compile on all supported SDKs.

Make the `linter` workflow do the checkout just like the others, without
explicitly specifying what to checkout.

Double the connection multiplexer timeout in tests, to avoid some tests
failing (Gears related tests) occasionally.