Rename existing blocking commands
The current naming for the blocking commands does not feel right. For
example "BzmPop", it feels like there are no naming rules. So rename
existing methods by this rule: Use capital letters at the beginning of
human-readable words, e.g. "Pop", and also use capital letters where the
Redis command has a single letter, e.g. "B", which stands for
"Blocking", "Z" which stands for "Set" and "M" which stands for "Multi".
So the new name is "BZMPop".

It's still a good time to do such renaming, because these commands have
not been released. Also it is a crucial moment to get the naming as
right as possible, for the same reason.